refactor: use extension code in same file
